Update data-collection-log-text.md
Sample showed creating the custom table with Computer and
FilePath, but these values were never populated using the
sample transformation. 

To add further confusion, the original description describing
the Compter and FilePath, implied that the columns would be
automatically populated.   I'm still not sure if they had to be included in the custom table in order to be populated in the source.